isu_info_bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.1.0
flake8==6.0.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
Jinja2==3.1.2
lxml==4.9.2
magic-filter==1.0.9
MarkupSafe==2.1.2
mccabe==0.7.0
multidict==6.0.4
Pillow==9.5.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dotenv==1.0.0
pytz==2022.7.1
redis==4.5.4
requests==2.28.2
SQLAlchemy==2.0.8
typing-extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2